requirements.txt
absl-py==2.1.0
alembic==1.14.1
attrs==25.1.0
blinker==1.8.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
distlib==0.3.9
easygui==0.98.3
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.1.24
fonttools==4.55.8
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
jax==0.4.13
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
Mako==1.3.9
MarkupSafe==2.1.5
matplotlib==3.7.5
mediapipe==0.10.11
ml-dtypes==0.2.0
numpy==1.24.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pickle-mixin==1.0.2
pillow==10.4.0
platformdirs==4.3.6
pocketsphinx==5.0.4
protobuf==3.20.3
PyAudio==0.2.14
pycparser==2.22
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sounddevice==0.5.1
SpeechRecognition==3.10.4
SQLAlchemy==2.0.38
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.31.2
Werkzeug==3.0.6
zipp==3.20.2
